+++ spec/ricordami/is_retrievable_spec.rb
@@ -33,10 +33,10 @@ describe Ricordami::IsRetrievable do
       Tenant.attribute :name
       instance = Tenant.new(:id => "hi")
       instance.save
-      Tenant.indices[:id].all.should == ["hi"]
+      Tenant.indices[:u_id].all.should == ["hi"]
       instance.name = "john"
       instance.save
-      Tenant.indices[:id].all.should == ["hi"]
+      Tenant.indices[:u_id].all.should == ["hi"]
     end
 
     it "returns all the instances with #all" do