+++ spec/ricordami/unique_index_spec.rb
@@ -13,7 +13,7 @@ describe Ricordami::UniqueIndex do
   it "is initialized with a model, a name and the fields to be unique" do
     @index.model.should == DataSource
     @index.fields.should == [:id]
-    @index.name.should == :id
+    @index.name.should == :u_id
   end
 
   it "can have a scope" do
@@ -23,7 +23,7 @@ describe Ricordami::UniqueIndex do
   end
 
   it "returns its internal index name with #uidx_key_name" do
-    @index.uidx_key_name.should == "DataSource:udx:id"
+    @index.uidx_key_name.should == "DataSource:udx:u_id"
   end
 
   it "returns its internal reference name with #ref_key_name" do
@@ -32,14 +32,14 @@ describe Ricordami::UniqueIndex do
 
   it "adds a string to the index with #add" do
     @index.add("ze-id", "allo")
-    Ricordami.redis.smembers("DataSource:udx:id").should == ["allo"]
+    Ricordami.redis.smembers("DataSource:udx:u_id").should == ["allo"]
   end
 
   it "also indices the hash index with #add if fields is not :id and :get_by is true" do
     DataSource.attribute :domain
     other = subject.new(DataSource, [:name, :domain], :get_by => true)
     other.add("ze-id", ["jobs", "apple.com"])
-    Ricordami.redis.smembers("DataSource:udx:name_domain").should == ["jobs_-::-_apple.com"]
+    Ricordami.redis.smembers("DataSource:udx:u_name_domain").should == ["jobs_-::-_apple.com"]
     Ricordami.redis.hget("DataSource:hsh:name_domain_to_id", "jobs_-::-_apple.com").should == "ze-id"
   end
 
@@ -68,12 +68,12 @@ describe Ricordami::UniqueIndex do
   it "removes a string from the index with #rem" do
     @index.add("ze-id", "allo")
     @index.rem("ze-id", "allo")
-    Ricordami.redis.smembers("DataSource:udx:id").should == []
+    Ricordami.redis.smembers("DataSource:udx:u_id").should == []
   end
 
   it "returns the redis command(s) to remove the value from the index when return_command is true" do
     command = @index.rem("ze-id", "allo", true)
-    command.should == [[:srem, ["DataSource:udx:id", "allo"]]]
+    command.should == [[:srem, ["DataSource:udx:u_id", "allo"]]]
   end
 
   it "returns the number of entries with #count" do